# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      success:
        type: boolean
    type: object
  model.Currency:
    properties:
      ADate:
        type: string
      code:
        type: string
      title:
        type: string
      value:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API service for work currency
  termsOfService: http://swagger.io/terms/
  title: KMF Currency API
  version: "1.0"
paths:
  /currency/{date}/{code}:
    get:
      description: get currency
      operationId: show-currency
      parameters:
      - description: date
        in: path
        name: date
        required: true
        type: string
      - description: 'code (default: \'
        in: path
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Currency'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: ShowCurrency
      tags:
      - currency
  /currency/save/{date}:
    get:
      description: save currency
      operationId: save-currency
      parameters:
      - description: date
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: SaveCurrency
      tags:
      - currency
swagger: "2.0"